requirements.txt
django-appconf

